% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthCurves.R
\docType{data}
\name{default.plate.96}
\alias{default.plate.96}
\title{Default labels for 96-well plates (character matrix with the same shape as the plate)}
\format{\preformatted{ chr [1:8, 1:12] "A1" "B1" "C1" "D1" "E1" "F1" "G1" "H1" ...
}}
\usage{
default.plate.96
}
\description{
Default labels for 96-well plates (character matrix with the same shape as the plate)
}
\keyword{datasets}

