% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logistic_growth_mle_normal.R
\name{logistic_growth_mle_norm}
\alias{logistic_growth_mle_norm}
\title{logistic.growth.mle.norm}
\usage{
logistic_growth_mle_norm(abundance, time, upper = 2 * max(abundance))
}
\arguments{
\item{abundance}{Abundance (eg. Optical density, counts etc).}

\item{time}{Numeric vector of time since start of measurement (currently does not use date format)}
}
\description{
Maximum likelihood estimation of logistic growth parameters assuming normal error
}
\examples{
data(navicula_growth)
fitted.readings <- with(navicula_growth,
                       logistic_growth_mle_norm(time=Time,
                                           abundance=ABS))
head(fitted.readings)
}
\keyword{growth}

