% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthCurves.R
\name{magellanTable}
\alias{magellanTable}
\title{Loads data from a .asc file exported by Magellan.}
\usage{
magellanTable(filePath, wellLabels = c(t(default.plate.96)),
  table.postprocess = validateTable)
}
\arguments{
\item{filePath}{path on the local file system to be passed to the parser}

\item{plateLabels}{a matrix of well labels that matches the shape of the plate
  data being analyzed.  Two default layouts are provided: \code{\link{default.plate.96}}
  and \code{\link{default.plate.384}}}
}
\value{
a data.frame object holding the parsed data.
}
\description{
Loads data from a .asc file exported by Magellan.
}

